epiphany/services/edit/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
boto3==1.34.139
numpy==1.26.4
Pillow==10.3.0
pytest==8.3.2
rembg==2.0.57
opencv-python-headless==4.10.0.84
requests==2.32.3
